feat: impove hadlers

+++ src/luminary/chat/application/usecases/command/delete_chat_use_case.py
@@ -1,3 +1,4 @@
+from common.domain.exceptions import InvariantViolationError
 from common.domain.value_objects.id import UserId
 
 from luminary.chat.application.interfaces.policies.chat_access_policy import (
@@ -25,6 +26,11 @@ class DeleteChatUseCase(IDeleteChatUseCase):
     async def execute(self, command: DeleteChatCommand) -> None:
         chat = await self.repository.get_by_id(ChatId(command.chat_id))
         self.access_policy.assert_is_allowed(UserId(command.user_id), chat)
+
+        if not chat.is_standalone():
+            raise InvariantViolationError(
+                "Only standalone chat can be deleted using this use case"
+            )
 
         chat.delete()
         await self.repository.save(chat)
